added selectors
