+++ project/src/store/app-state/selectors.js
@@ -1,0 +1,7 @@
+export const getIsDataLoaded = ({STATE}) => STATE.isDataLoaded;
+
+export const getIsCommentLoaded = ({STATE}) => STATE.isCommentLoaded;
+
+export const getComment = ({STATE}) => STATE.comment;
+
+export const getRating = ({STATE}) => STATE.rating;
