+++ project/src/store/app-user/selectors.js
@@ -1,0 +1,6 @@
+export const getUser = ({USER}) => USER.email;
+
+export const getAvatarUrl = ({USER}) => USER.avatarUrl;
+
+export const getAuthorizationStatus = ({USER}) => USER.authorizationStatus;
+
